Msdata.SystEnumFunctionalTypes (mscmp_syst_enums v0.1.0)

The data structure defining the functional types associated with a given enumeration.

Functional types allow the application behavior associated with list choices to be decoupled from the information conveyance that the same list of choices may be able to provide. For example, the application may only need to recognize the difference between 'active' and 'inactive' order records as a status, however a user may want to see some reflection of why an order was made inactive, such as it was cancelled or that it was fulfilled. In this case would create two entries of functional type inactive, one each for cancelled and fulfilled. Functionally the application will see the same value, but user reporting can reflect the nuance not needed by the application functionality.

Defined in MscmpSystEnums.

Types

t()

@type t() :: %Msdata.SystEnumFunctionalTypes{
  __meta__: Ecto.Schema.Metadata.t(),
  diag_role_created: String.t() | nil,
  diag_role_modified: String.t() | nil,
  diag_row_version: integer() | nil,
  diag_timestamp_created: DateTime.t() | nil,
  diag_timestamp_modified: DateTime.t() | nil,
  diag_update_count: integer() | nil,
  diag_wallclock_modified: DateTime.t() | nil,
  display_name: String.t() | nil,
  enum: Msdata.SystEnums.t() | Ecto.Association.NotLoaded.t() | nil,
  enum_id: Ecto.UUID.t() | nil,
  external_name: String.t() | nil,
  id: Ecto.UUID.t() | nil,
  internal_name: MscmpSystEnums.Types.enum_functional_type_name() | nil,
  syst_defined: boolean() | nil,
  syst_description: String.t() | nil,
  user_description: String.t() | nil
}